requirements.txt
blinker==1.4
Click==7.0
Flask==1.1.1
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
flask-marshmallow==0.10.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
marshmallow==3.4.0
marshmallow-sqlalchemy==0.22.2
pkg-resources==0.0.0
PyJWT==1.7.1
python-dotenv==0.11.0
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==1.0.0
